pyproject.toml
[tool.poetry]
name = "xsb-fluc"
version = "0.1.0"
description = "X-SP fluctuation code used in S. Dupourqué's PhD thesis"
authors = ["sdupourque <sdupourque@irap.omp.eu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
numpy = "^1.26.1"
scipy = "^1.11.3"
jax = "^0.4.20"
jaxlib = "^0.4.20"
dm-haiku = "^0.0.10"
numpyro = "^0.13.2"
astropy = "^6.0.0"
regions = "^0.8"
matplotlib = "^3.8.2"
cmasher = "^1.7.2"
reproject = "^0.13.0"
jupyterlab = "^4.0.11"
notebook = "^7.0.7"
ipywidgets = "^8.1.1"
arviz = "^0.17.0"
numba = "^0.58.1"
vorbin = "^3.1.5"


[tool.poetry.group.docs.dependencies]
mkdocs = "^1.5.3"
mkdocs-material = "^9.4.8"
mkdocstrings = {extras = ["python"], version = "^0.23.0"}
mkdocs-jupyter = "^0.24.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
